Major bugfixes in liveRec: * StandaloneMultiEventTypeLabeler.m: NaN can't be assigned to int (but matlab doesn't throw an error) * idScoresBAC.m: groundtruth was setup wrongly (only blocks completely enclosed by events were used)

+++ src/evaluation/liveRec/StandaloneMultiEventTypeLabeler.m
@@ -18,7 +18,7 @@ classdef StandaloneMultiEventTypeLabeler < LabelCreators.MultiEventTypeLabeler
         %% -------------------------------------------------------------------------------
         function y = labelBlock( obj, blockAnnotations )
             [activeTypes, relBlockEventOverlap, ~] = obj.getActiveTypes( blockAnnotations );
-            y = int16(activeTypes);
+            y = single(activeTypes);
             if strcmp( obj.negOut, 'rest' )
                 y(~activeTypes & relBlockEventOverlap > obj.maxNegBlockToEventRatio) = NaN;
             end
